Python_Bot/requirements.txt
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloud-init==23.3.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
dnspython==2.1.0
Flask==3.0.0
gpg==1.16.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Markdown==3.3.6
MarkupSafe==2.1.3
more-itertools==8.10.0
MouseInfo==0.1.3
netifaces==0.11.0
oauthlib==3.2.0
pexpect==4.8.0
Pillow==10.1.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyAutoGUI==0.9.54
pygame==2.5.2
PyGetWindow==0.0.9
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
PyMsgBox==1.0.9
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.30
pyserial==3.5
python-apt==2.4.0+ubuntu2
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-magic==0.4.24
python3-xlib==0.15
pyttsx3==2.90
pytube==15.0.0
pytweening==1.0.7
pytz==2022.1
pywhatkit==5.4
PyYAML==5.4.1
requests==2.31.0
requests-toolbelt==0.9.1
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sos==4.5.6
soupsieve==2.5
SpeechRecognition==3.10.0
ssh-import-id==5.11
systemd-python==234
Twisted==22.1.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
Werkzeug==3.0.1
wikipedia==1.4.0
xkit==0.0.0
zipp==1.0.0
zope.interface==5.4.0
